feat(avo-1899): allow teacher to view assignment response edit as pupil
