+++ src/assignment/assignment.gql.ts
@@ -125,6 +125,9 @@ export const GET_ASSIGNMENTS_BY_OWNER_ID = gql`
 			owner_profile_id
 			updated_at
 			created_at
+			owner {
+				full_name
+			}
 		}
 		count: app_assignments_v2_aggregate(
 			where: {
@@ -195,6 +198,9 @@ export const GET_ASSIGNMENTS_BY_RESPONSE_OWNER_ID = gql`
 					or_id
 				}
 				id
+			}
+			owner {
+				full_name
 			}
 		}
 		count: app_assignments_v2_aggregate(
@@ -388,6 +394,9 @@ export const GET_ASSIGNMENT_WITH_RESPONSE = gql`
 					first_name
 					last_name
 				}
+			}
+			owner {
+				full_name
 			}
 		}
 	}
