+++ src/assignment/assignment.routes.tsx
@@ -6,7 +6,7 @@ import { APP_PATH } from '../constants';
 import AssignmentCreate from './views/AssignmentCreate';
 import AssignmentDetail from './views/AssignmentDetail';
 import AssignmentEdit from './views/AssignmentEdit';
-import AssignmentResponseEdit from './views/AssignmentResponseEdit/AssignmentResponseEdit';
+import AssignmentResponseEditPage from './views/AssignmentResponseEdit/AssignmentResponseEditPage';
 import AssignmentResponses from './views/AssignmentResponses';
 
 export const renderAssignmentRoutes = (): ReactNode[] => [
@@ -29,13 +29,13 @@ export const renderAssignmentRoutes = (): ReactNode[] => [
 		key={APP_PATH.ASSIGNMENT_RESPONSES.route}
 	/>,
 	<SecuredRoute
-		component={AssignmentResponseEdit}
+		component={AssignmentResponseEditPage}
 		exact
 		path={APP_PATH.ASSIGNMENT_RESPONSE_CREATE.route}
 		key={APP_PATH.ASSIGNMENT_RESPONSE_CREATE.route}
 	/>,
 	<SecuredRoute
-		component={AssignmentResponseEdit}
+		component={AssignmentResponseEditPage}
 		exact
 		path={APP_PATH.ASSIGNMENT_RESPONSE_EDIT.route}
 		key={APP_PATH.ASSIGNMENT_RESPONSE_EDIT.route}
