+++ src/assignment/views/AssignmentResponseEdit/tabs/AssignmentResponseAssignmentTab.tsx
@@ -1,4 +1,4 @@
-import { Container, Flex, Spacer, Spinner } from '@viaa/avo2-components';
+import { Container } from '@viaa/avo2-components';
 import { Avo } from '@viaa/avo2-types';
 import React, { FunctionComponent } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -7,40 +7,17 @@ import { ErrorView } from '../../../../error/views';
 import BlockList from '../../../../shared/components/BlockList/BlockList';
 
 interface AssignmentResponseAssignmentTabProps {
-	assignment: Avo.Assignment.Assignment_v2 | null;
-	assignmentLoading: boolean;
-	assignmentError: any | null;
+	assignment: Avo.Assignment.Assignment_v2;
 }
 
 const AssignmentResponseAssignmentTab: FunctionComponent<AssignmentResponseAssignmentTabProps> = ({
 	assignment,
-	assignmentError,
-	assignmentLoading,
 }) => {
 	const [t] = useTranslation();
 
 	// Render
 
 	const renderAssignmentBlocks = () => {
-		if (assignmentLoading) {
-			return (
-				<Spacer margin="top-extra-large">
-					<Flex orientation="horizontal" center>
-						<Spinner size="large" />
-					</Flex>
-				</Spacer>
-			);
-		}
-		if (assignmentError) {
-			return (
-				<ErrorView
-					message={t(
-						'assignment/views/assignment-response-edit___het-ophalen-van-de-opdracht-is-mislukt'
-					)}
-					icon="alert-triangle"
-				/>
-			);
-		}
 		if ((assignment?.blocks?.length || 0) === 0) {
 			return (
 				<ErrorView
